adding functionnality delete and update for artist and artwork and adding 10 pictures

+++ server/database/models/UserRepository.js
@@ -72,7 +72,7 @@ class UserRepository extends AbstractRepository {
 
   async delete(id) {
     const [rows] = await this.database.query(
-      `DELETE from ${this.table} WHERE id=?`,
+      `DELETE from ${this.table} WHERE id=? AND isadmin!=1`,
       [id]
     );
